[ext-msg] factoring EmbedlyClient out of S3ScreenshotStore
