+++ kifi-backend/modules/eliza/app/com/keepit/dev/ElizaDevModule.scala
@@ -1,7 +1,7 @@
 package com.keepit.dev
 
 import com.keepit.inject.CommonDevModule
-import com.keepit.eliza.{DevElizaMailNotificationsModule, ElizaModule}
+import com.keepit.eliza.{DevElizaExternalEmailModule, ElizaModule}
 import com.keepit.realtime.ElizaUrbanAirshipModule
 import com.keepit.common.cache.HashMapMemoryCacheModule
 import com.keepit.common.cache.ElizaCacheModule
@@ -13,6 +13,6 @@ case class ElizaDevModule() extends ElizaModule(
   urbanAirshipModule = ElizaUrbanAirshipModule(),
   storeModule = ElizaDevStoreModule()
 ) with CommonDevModule {
-  val elizaMailSettingsModule = DevElizaMailNotificationsModule()
+  val elizaMailSettingsModule = DevElizaExternalEmailModule()
 }
